requirements.txt
click==8.0.1
requests==2.27.1
dataclasses==0.6
environs==9.3.2
Flask==2.0.1
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.2
pydevd-pycharm==212.4746.13
python-dotenv==0.18.0
PyYAML==5.3.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
pyrabbit==1.1.0
sendgrid==6.10.0
discordwebhook==1.0.3
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.6
Flask-BasicAuth==0.2.0